Many bugfixes/ux changes
Set up enemy spawn rates and movespeeds to be able to scale with (as yet
unimplemented) difficulty. Added many display assets, as well as
implemented a do-nothing settings page. Added sound control options,
implemented in execution, unable to set them as the player as yet. All
of this is done from the new title screen, which plays a theme (somewhat
cheerier than the main theme).